// algo-2022-10/dijkstra-inno-beer-2023-08/app-dev/src/game/Graph.types.ts
export interface AbstractGraph {
    edges: Edge[];
    vertices: Vertex[];
    cheapestPath: number[];
    curVertexIndex: number;
}
export const defaultAbstractGraph: AbstractGraph = {
    edges: [],
    vertices: [],
    cheapestPath: [],
    curVertexIndex: -1
};

export interface Edge {
    vertex0: number;
    vertex1: number;
    cost: EdgeCost;
}

export interface Vertex {
    processed: boolean;
    accessCost: number;
    edgeIndex: number;
}

const UNDEFINED_INDEX = -1;
export const UNDEFINED_COST = -1;
export const defaultVertex: Vertex = {
    processed: false,
    accessCost: UNDEFINED_COST,
    edgeIndex: UNDEFINED_INDEX
};

export interface EdgeCost {
    cost: number;
    v0v1Cost: number;
    v1v0Cost: number;
}